feat: Update devDependencies, adding @types/node package
This change updates the devDependencies in the package.json file by
adding the "@types/node" package with a minimum version of "^18". This
is done to ensure the TypeScript type definitions for Node.js are up to
date. It will help improve type safety and provide better autocompletion
and error checking when working with Node.js libraries and modules in
the codebase.

Closes #123