docs: Update package dependencies to include next-themes (#1234)
Update the package dependencies to include next-themes version 0.2.1.
The previous version, nextra-theme-blog 2.13.2, had a dependency on
next-themes as well, but it is now unnecessary. By directly including
next-themes, we reduce the complexity of the dependency tree and
streamline our codebase. This change helps improve the maintainability
and performance of our application. Refactor code to remove unnecessary
nextra-theme-blog dependency.

Fixes #1234.